requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
pytesseract>=0.3.10
Pillow>=10.3.0
opencv-python>=4.9.0.80
numpy>=1.26.4
pdf2image>=1.17.0
pydantic>=2.7.1
pandas>=2.2.2
tiktoken>=0.7.0
# Optional preferred OCR; pipeline tries to import gracefully
dots-ocr; python_version >= "3.8"
